code fix after review
