+++ src/constants/apiConstants/apiConstants.ts
@@ -1,0 +1,20 @@
+export const API_BASE_URL = "https://api.jikan.moe/v4/anime";
+export const THUMBNAIL =
+    "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRBNnH7hPrYmvHPtcImonVIDwGgMIgtg5URew&s";
+export const ONGOING = "ongoing";
+export const UNKNOWN_YEAR = "Unknown";
+export const DEFAULT_DESCRIPTION = "Wops, description is not found now !";
+export const MONTHS = [
+    "January",
+    "February",
+    "March",
+    "April",
+    "May",
+    "June",
+    "July",
+    "August",
+    "September",
+    "October",
+    "November",
+    "December",
+];
